+++ packages/widgetbook/lib/src/core/accessibility/accessibility.dart
@@ -1,0 +1,2 @@
+export 'accessibility_config.dart';
+export 'guidelines.dart';